% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{affine_invariant_distance}
\alias{affine_invariant_distance}
\title{Affine-Invariant Distance}
\usage{
affine_invariant_distance(mu1, Sigma1, mu2, Sigma2)
}
\arguments{
\item{mu1}{Mean vector of the first distribution}

\item{Sigma1}{Covariance matrix of the first distribution}

\item{mu2}{Mean vector of the second distribution}

\item{Sigma2}{Covariance matrix of the second distribution}
}
\value{
The affine-invariant distance (non-negative scalar)
}
\description{
Computes the affine-invariant Riemannian distance between two multivariate
normal distributions. Uses the log-Euclidean metric on covariance matrices.
}
\references{
Pennec, X., Fillard, P., & Ayache, N. (2006). A Riemannian framework
for tensor computing. International Journal of Computer Vision, 66(1), 41-66.
}
