% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{bhattacharyya_distance}
\alias{bhattacharyya_distance}
\title{Bhattacharyya Distance Between Multivariate Normals}
\usage{
bhattacharyya_distance(mu1, Sigma1, mu2, Sigma2)
}
\arguments{
\item{mu1}{Mean vector of the first distribution}

\item{Sigma1}{Covariance matrix of the first distribution}

\item{mu2}{Mean vector of the second distribution}

\item{Sigma2}{Covariance matrix of the second distribution}
}
\value{
The Bhattacharyya distance (non-negative scalar)
}
\description{
Computes the Bhattacharyya distance between two multivariate normal distributions.
This distance is related to the Bhattacharyya coefficient and Bayes classification error.
}
\examples{
mu1 <- c(0, 0)
Sigma1 <- diag(2)
mu2 <- c(1, 1)
Sigma2 <- matrix(c(2, 0.5, 0.5, 1), 2, 2)
bhattacharyya_distance(mu1, Sigma1, mu2, Sigma2)

}
