% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{symmetrized_kl}
\alias{symmetrized_kl}
\title{Symmetrized KL Divergence}
\usage{
symmetrized_kl(mu1, Sigma1, mu2, Sigma2)
}
\arguments{
\item{mu1}{Mean vector of the first distribution}

\item{Sigma1}{Covariance matrix of the first distribution}

\item{mu2}{Mean vector of the second distribution}

\item{Sigma2}{Covariance matrix of the second distribution}
}
\value{
The symmetrized KL divergence (non-negative scalar)
}
\description{
Computes the symmetrized Kullback-Leibler divergence, which is the average
of KL(P||Q) and KL(Q||P). This provides a symmetric measure.
}
